# Optimise get video event tracking for minimal memory allocs