# Description

 85% improvement in cpu usage and 25% improvement in memory usage of `GetVideoEventTracking()`

```
 // Original:
 // Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkGetVideoEventTracking$ github.com/PubMatic-OpenWrap/prebid-server/endpoints/events

 // goos: linux
 // goarch: arm64
 // pkg: github.com/PubMatic-OpenWrap/prebid-server/endpoints/events
 // BenchmarkGetVideoEventTracking-8   	   19048	     78882 ns/op	   31590 B/op	     128 allocs/op
 // BenchmarkGetVideoEventTracking-8   	   27333	     40491 ns/op	   31589 B/op	     128 allocs/op
 // BenchmarkGetVideoEventTracking-8   	   28392	     45111 ns/op	   31586 B/op	     128 allocs/op
 // BenchmarkGetVideoEventTracking-8   	   18160	     83581 ns/op	   31585 B/op	     128 allocs/op
 // BenchmarkGetVideoEventTracking-8   	   16633	     77993 ns/op	   31591 B/op	     128 allocs/op
 // PASS
 // ok  	github.com/PubMatic-OpenWrap/prebid-server/endpoints/events	1.807s

 // Refactored-GetVideoEventTracking-using-replaceMacros:
 // BenchmarkGetVideoEventTracking-8   	   65928	     16866 ns/op	   10434 B/op	      96 allocs/op
 // BenchmarkGetVideoEventTracking-8   	   66710	     18611 ns/op	   10433 B/op	      96 allocs/op
 // BenchmarkGetVideoEventTracking-8   	   66448	     17244 ns/op	   10433 B/op	      96 allocs/op
 // BenchmarkGetVideoEventTracking-8   	   35112	     38085 ns/op	   10433 B/op	      96 allocs/op
 // BenchmarkGetVideoEventTracking-8   	   40941	     27584 ns/op	   10434 B/op	      96 allocs/op
```
# Checklist:

- [ ] PR commit list is unique (rebase/pull with the origin branch to keep master clean).
- [ ] JIRA number is added in the PR title and the commit message.
- [ ] Updated the `header-bidding` repo with appropiate commit id.
- [ ] Documented the new changes.

For Prebid upgrade, refer: https://inside.pubmatic.com:8443/confluence/display/Products/Prebid-server+upgrade